Meta: rename features
