+++ source/features/extend-navigation-shortcuts.tsx
@@ -55,10 +55,6 @@ function init(): void {
 }
 
 void features.add(import.meta.url, {
-	shortcuts: {
-		j: 'Focus the comment/file below',
-		k: 'Focus the comment/file above',
-	},
 	include: [
 		pageDetect.hasComments,
 	],
