+++ source/options-storage.ts
@@ -41,7 +41,24 @@ export const renamedFeatures = new Map<string, string>([
 	['quick-pr-diff-options', 'one-click-diff-options'],
 	['quick-review-buttons', 'one-click-review-submission'],
 	['wait-for-build', 'wait-for-checks'],
-	['pull-request-hotkey', 'pull-request-hotkeys'],
+	['pull-request-hotkey', 'pull-request-shortcuts'],
+	['keyboard-navigation', 'extend-navigation-shortcuts'],
+	['pagination-hotkey', 'pagination-shortcuts'],
+	['profile-hotkey', 'profile-shortcut'],
+	['show-associated-branch-prs-on-fork', 'fork-branches-associated-prs'],
+	['copy-on-y', 'copy-permalink-shortcut'],
+	['show-user-top-repositories', 'user-top-repositories-link'],
+	['show-whitespace', 'render-whitespace'],
+	['swap-branches-on-compare', 'swap-compared-branches-link'],
+	['link-to-github-io', 'github-io-link'],
+	['show-names', 'show-user-names'],
+	['tags-on-commits-list', 'commit-associated-tags'],
+	['same-branch-author-commits', 'authored-commits-same-view'],
+	['link-to-changelog-file', 'changelog-file-link'],
+	['conversation-links-on-repo-lists', 'repo-conversation-links'],
+	['link-to-prior-blame-line', 'blame-prior-change-same-line'],
+	['linkify-labels-on-dashboard', 'linkify-dashboard-labels'],
+	['same-page-definition-jump', 'quick-go-to-definition'],
 ]);
 
 export function getNewFeatureName(possibleFeatureName: string): FeatureID | undefined {
