+++ source/refined-github.ts
@@ -48,8 +48,8 @@ import './features/selection-in-new-tab';
 import './features/quick-comment-hiding';
 import './features/quick-comment-edit';
 import './features/open-all-notifications';
-import './features/copy-on-y';
-import './features/profile-hotkey';
+import './features/copy-permalink-shortcut';
+import './features/profile-shortcut';
 import './features/close-out-of-view-modals';
 import './features/improve-shortcut-help';
 import './features/copy-file';
@@ -60,11 +60,11 @@ import './features/linkify-code';
 import './features/download-folder-button';
 import './features/linkify-branch-references';
 import './features/open-all-conversations';
-import './features/pagination-hotkey';
-import './features/conversation-links-on-repo-lists';
+import './features/pagination-shortcuts';
+import './features/repo-conversation-links';
 import './features/global-conversation-list-filters';
 import './features/more-conversation-filters';
-import './features/sort-conversations-by-update-time'; // Must be after global-conversation-list-filters and more-conversation-filters and conversation-links-on-repo-lists
+import './features/sort-conversations-by-update-time'; // Must be after `global-conversation-list-filters` and `more-conversation-filters` and `conversation-links-on-repo-lists`
 import './features/pinned-issues-update-time';
 import './features/latest-tag-button';
 import './features/default-branch-button';
@@ -75,7 +75,7 @@ import './features/sync-pr-commit-title';
 import './features/open-ci-details-in-new-tab';
 import './features/wait-for-checks';
 import './features/hide-inactive-deployments';
-import './features/pull-request-hotkeys';
+import './features/pull-request-shortcuts';
 import './features/one-click-review-submission';
 import './features/embed-gist-inline';
 import './features/comments-time-machine-links';
@@ -85,14 +85,14 @@ import './features/create-release-shortcut';
 import './features/patch-diff-links';
 import './features/parse-backticks';
 import './features/mark-merge-commits-in-list';
-import './features/swap-branches-on-compare';
+import './features/swap-compared-branches-link';
 import './features/reactions-avatars';
-import './features/show-names';
+import './features/show-user-names';
 import './features/previous-next-commit-buttons';
 import './features/preserve-whitespace-option-in-nav';
 import './features/extend-diff-expander';
 import './features/profile-gists-link';
-import './features/show-user-top-repositories';
+import './features/user-top-repositories';
 import './features/set-default-repositories-type-to-sources';
 import './features/user-profile-follower-badge';
 import './features/highlight-non-default-base-branch';
@@ -107,8 +107,8 @@ import './features/fit-textareas';
 import './features/collapsible-content-button';
 import './features/resolve-conflicts';
 import './features/follow-file-renames';
-import './features/enable-file-links-in-compare-view'; // Must be before more-file-links
-import './features/more-file-links';
+import './features/enable-file-links-in-compare-view';
+import './features/more-file-links'; // Must be after `enable-file-links-in-compare-view`
 import './features/tags-dropdown';
 import './features/pr-filters';
 import './features/quick-file-edit';
@@ -123,24 +123,24 @@ import './features/toggle-everything-with-alt';
 import './features/suggest-commit-title-limit';
 import './features/highest-rated-comment';
 import './features/clean-conversation-filters';
-import './features/tags-on-commits-list';
+import './features/commit-associated-tags';
 import './features/forked-to';
 import './features/edit-readme';
 import './features/list-prs-for-file';
 import './features/pr-branch-auto-delete';
-import './features/linkify-symbolic-links'; // Must be before show-whitespace
-import './features/show-whitespace';
+import './features/linkify-symbolic-links';
+import './features/render-whitespace'; // Must be after `linkify-symbolic-links`
 import './features/restore-file';
 import './features/hidden-review-comments-indicator';
 import './features/reload-failed-proxied-images';
 import './features/highlight-collaborators-and-own-conversations';
 import './features/embed-gist-via-iframe';
 import './features/one-click-pr-or-gist';
-import './features/link-to-prior-blame-line';
+import './features/blame-prior-change-same-line';
 import './features/dim-bots';
 import './features/conflict-marker';
 import './features/html-preview-link';
-import './features/linkify-labels-on-dashboard';
+import './features/linkify-dashboard-labels';
 import './features/linkify-user-location';
 import './features/repo-age';
 import './features/user-local-time';
@@ -162,11 +162,11 @@ import './features/linkify-notification-repository-header';
 import './features/stop-redirecting-in-notification-bar';
 import './features/prevent-link-loss';
 import './features/first-published-tag-for-merged-pr';
-import './features/show-associated-branch-prs-on-fork';
+import './features/fork-branches-associated-prs';
 import './features/quick-review';
 import './features/fork-source-link-same-view';
 import './features/pr-jump-to-first-non-viewed-file';
-import './features/keyboard-navigation';
+import './features/extend-navigation-shortcuts';
 import './features/vertical-front-matter';
 import './features/use-first-commit-message-for-new-prs';
 import './features/linkify-user-edit-history-popup';
@@ -177,10 +177,10 @@ import './features/clean-conversation-headers';
 import './features/stop-pjax-loading-with-esc';
 import './features/highlight-deleted-and-added-files-in-diffs';
 import './features/convert-release-to-draft';
-import './features/same-page-definition-jump';
+import './features/quick-go-to-definition';
 import './features/new-repo-disable-projects-and-wikis';
 import './features/table-input';
-import './features/link-to-github-io';
+import './features/github-io-link';
 import './features/scheduled-and-manual-workflow-indicators';
 import './features/convert-pr-to-draft-improvements';
 import './features/git-checkout-pr';
@@ -193,7 +193,7 @@ import './features/clean-repo-sidebar';
 import './features/rgh-feature-descriptions';
 import './features/wait-for-attachments';
 import './features/useful-forks';
-import './features/link-to-changelog-file';
+import './features/changelog-file-link';
 import './features/rgh-sponsor-button';
 import './features/rgh-linkify-features';
 import './features/conversation-activity-filter';
@@ -210,10 +210,10 @@ import './features/comment-on-draft-pr-indicator';
 import './features/select-notifications';
 import './features/clean-repo-tabs';
 import './features/rgh-welcome-issue';
-import './features/same-branch-author-commits';
+import './features/authored-commits-same-view';
 import './features/prevent-pr-merge-panel-opening';
 import './features/rgh-improve-new-issue-form';
 import './features/easy-toggle-commit-messages';
 import './features/command-palette-navigation-shortcuts';
-import './features/link-to-compare-diff';
+import './features/compare-diff-link';
 import './features/hide-low-quality-comments';
